Extract test helpers
